# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API para el sistema de monitoreo de incidentes de inseguridad turística en Cartagena
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  correo: string;
  contrasena: string;
}

export interface RegisterInput {
  nombre: string;
  correo: string;
  contrasena: string;
}

export type UsuarioRol = typeof UsuarioRol[keyof typeof UsuarioRol];


export const UsuarioRol = {
  administrador: 'administrador',
  turista: 'turista',
} as const;

export interface Usuario {
  _id: string;
  nombre: string;
  correo: string;
  rol: UsuarioRol;
  /** @nullable */
  fechaRegistro?: string | null;
}

export interface AuthResponse {
  usuario: Usuario;
  token: string;
}

export type UsuarioUpdateRol = typeof UsuarioUpdateRol[keyof typeof UsuarioUpdateRol];


export const UsuarioUpdateRol = {
  administrador: 'administrador',
  turista: 'turista',
} as const;

export interface UsuarioUpdate {
  nombre?: string;
  correo?: string;
  rol?: UsuarioUpdateRol;
}

export interface Incidente {
  _id: string;
  /** @nullable */
  tipoId?: string | null;
  /** @nullable */
  usuarioId?: string | null;
  /** @nullable */
  zonaId?: string | null;
  descripcion: string;
  fecha: string;
  /** @nullable */
  ubicacion?: string | null;
  etiquetas?: string[];
  /** @nullable */
  tipoNombre?: string | null;
  /** @nullable */
  zonaNombre?: string | null;
  /** @nullable */
  usuarioNombre?: string | null;
}

export interface IncidenteInput {
  tipoId?: string;
  zonaId?: string;
  descripcion: string;
  ubicacion?: string;
  etiquetas?: string[];
}

export interface IncidenteUpdate {
  tipoId?: string;
  zonaId?: string;
  descripcion?: string;
  ubicacion?: string;
  etiquetas?: string[];
}

export interface IncidenteListResponse {
  data: Incidente[];
  total: number;
  page: number;
  limit: number;
}

export interface ZonaTuristica {
  _id: string;
  nombre: string;
  /** @nullable */
  ciudad?: string | null;
  /** @nullable */
  nivelRiesgo?: string | null;
}

export interface TipoIncidente {
  _id: string;
  nombre: string;
  /** @nullable */
  descripcion?: string | null;
}

export interface DashboardStats {
  totalIncidentes: number;
  totalUsuarios: number;
  totalZonas: number;
  incidentesHoy: number;
  incidentesEsteMes?: number;
}

export interface ZonaRiesgoStat {
  zonaId: string;
  zonaNombre: string;
  total: number;
  /** @nullable */
  nivelRiesgo?: string | null;
}

export interface TipoFrecuenciaStat {
  tipoId: string;
  tipoNombre: string;
  total: number;
}

export interface IncidenteMesStat {
  mes: string;
  total: number;
  anio?: number;
}

export interface ReporteroStat {
  usuarioId: string;
  usuarioNombre: string;
  total: number;
  /** @nullable */
  correo?: string | null;
}

export type ListIncidentesParams = {
page?: number;
limit?: number;
zona?: string;
tipo?: string;
fechaDesde?: string;
fechaHasta?: string;
};