# lib/api-zod/src/generated/types/incidenteUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API para el sistema de monitoreo de incidentes de inseguridad turística en Cartagena
 * OpenAPI spec version: 0.1.0
 */

export interface IncidenteUpdate {
  tipoId?: string;
  zonaId?: string;
  descripcion?: string;
  ubicacion?: string;
  etiquetas?: string[];
}