# lib/api-zod/src/generated/types/zonaRiesgoStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API para el sistema de monitoreo de incidentes de inseguridad turística en Cartagena
 * OpenAPI spec version: 0.1.0
 */

export interface ZonaRiesgoStat {
  zonaId: string;
  zonaNombre: string;
  total: number;
  /** @nullable */
  nivelRiesgo?: string | null;
}