# lib/api-zod/src/src/generated/types/incidenteListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API para el sistema de monitoreo de incidentes de inseguridad turística en Cartagena
 * OpenAPI spec version: 0.1.0
 */
import type { Incidente } from './incidente';

export interface IncidenteListResponse {
  data: Incidente[];
  total: number;
  page: number;
  limit: number;
}